feat(pricing): refactor pricing page and add features comparison component
